packages added and ProductController and GenericEntity moved to correct packages
